pyproject.toml
[project]
name = "autom8"
version = "0.1.0"
description = "Minimal terminal-based coding agent using OpenAI tool calling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyfiglet",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
py-modules = ["autom8", "agent", "tools"]

[project.scripts]
"8" = "autom8:main"